refactor: move `ark-ff` based `FieldElement` to a new sub-crate
