+++ Cargo.toml
@@ -22,6 +22,7 @@ members = [
     "starknet-crypto",
     "starknet-signers",
     "starknet-accounts",
+    "starknet-ff",
 ]
 
 [dependencies]
