+++ starknet-crypto/Cargo.toml
@@ -13,7 +13,7 @@ Low-level cryptography utilities for StarkNet
 keywords = ["ethereum", "starknet", "web3"]
 
 [dependencies]
-ark-ff = "0.3.0"
+starknet-ff = { version = "0.1.0", path = "../starknet-ff" }
 # Parity's scale codec locks us here
 bitvec = "0.20.4"
 crypto-bigint = "0.3.2"
