+++ starknet-crypto/src/test_utils.rs
@@ -1,6 +1,6 @@
 #![cfg(test)]
 
-use crate::field_element::FieldElement;
+use crate::FieldElement;
 
 pub fn field_element_from_be_hex(hex: &str) -> FieldElement {
     let decoded = hex::decode(hex.trim_start_matches("0x")).unwrap();
