+++ starknet-ff/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "starknet-ff"
+version = "0.1.0"
+authors = ["Jonathan LEI <me@xjonathan.dev>"]
+license = "MIT OR Apache-2.0"
+edition = "2021"
+readme = "README.md"
+repository = "https://github.com/xJonathanLEI/starknet-rs"
+homepage = "https://starknet.rs/"
+description = """
+StarkNet field element type
+"""
+keywords = ["ethereum", "starknet", "web3"]
+
+[dependencies]
+ark-ff = "0.3.0"
+# Parity's scale codec locks us here
+bitvec = "0.20.4"
